Move `PartialMerkleTree` to the `block` module
