+++ lib/block/PartialMerkleTree.js
