+++ lib/block/merkleblock.js
@@ -10,7 +10,7 @@ var BufferReader = require('../encoding/bufferreader');
 var BufferWriter = require('../encoding/bufferwriter');
 var Hash = require('../crypto/hash');
 var Transaction = require('../transaction');
-var PartialMerkleTree = require('../deterministicmnlist/PartialMerkleTree');
+var PartialMerkleTree = require('./PartialMerkleTree');
 var $ = require('../util/preconditions');
 
 /**
