+++ lib/deterministicmnlist/SimplifiedMNListDiff.js
@@ -9,7 +9,7 @@ var _ = require('lodash');
 var isHexString = require('../util/js').isHexa;
 
 var SimplifiedMNListEntry = require('./SimplifiedMNListEntry');
-var PartialMerkleTree = require('./PartialMerkleTree');
+var PartialMerkleTree = require('../block/PartialMerkleTree');
 var Transaction = require('../transaction');
 var constants = require('../constants');
